+++ localflavor/us/models.py
@@ -20,7 +20,6 @@ class USStateField(CharField):
     def deconstruct(self):
         name, path, args, kwargs = super(USStateField, self).deconstruct()
         del kwargs['choices']
-        del kwargs['max_length']
         return name, path, args, kwargs
 
 
@@ -45,7 +44,6 @@ class USPostalCodeField(CharField):
     def deconstruct(self):
         name, path, args, kwargs = super(USPostalCodeField, self).deconstruct()
         del kwargs['choices']
-        del kwargs['max_length']
         return name, path, args, kwargs
 
 
@@ -69,11 +67,6 @@ class USZipCodeField(CharField):
         kwargs['max_length'] = 10
         super(USZipCodeField, self).__init__(*args, **kwargs)
 
-    def deconstruct(self):
-        name, path, args, kwargs = super(USZipCodeField, self).deconstruct()
-        del kwargs['max_length']
-        return name, path, args, kwargs
-
     def formfield(self, **kwargs):
         defaults = {'form_class': forms.USZipCodeField}
         defaults.update(kwargs)
@@ -90,11 +83,6 @@ class PhoneNumberField(CharField):
     def __init__(self, *args, **kwargs):
         kwargs['max_length'] = 20
         super(PhoneNumberField, self).__init__(*args, **kwargs)
-
-    def deconstruct(self):
-        name, path, args, kwargs = super(PhoneNumberField, self).deconstruct()
-        del kwargs['max_length']
-        return name, path, args, kwargs
 
     def formfield(self, **kwargs):
         from localflavor.us.forms import USPhoneNumberField
@@ -115,11 +103,6 @@ class USSocialSecurityNumberField(CharField):
     def __init__(self, *args, **kwargs):
         kwargs['max_length'] = 11
         super(USSocialSecurityNumberField, self).__init__(*args, **kwargs)
-
-    def deconstruct(self):
-        name, path, args, kwargs = super(USSocialSecurityNumberField, self).deconstruct()
-        del kwargs['max_length']
-        return name, path, args, kwargs
 
     def formfield(self, **kwargs):
         from localflavor.us.forms import (USSocialSecurityNumberField as